Revert "Revert "fix ui chat""
This reverts commit ad09a3f09be8c5072478d5d38080cc8cdbc8e432.

+++ app/src/main/java/com/example/dattingapp/Models/Filter.java
@@ -4,5 +4,5 @@ public class Filter {
     public  int minAge;
     public  int maxAge;
     public  String gender;
-    public  int distance;
+    public  Double distance;
 }
